+++ cmake/cora.cmake
@@ -9,6 +9,7 @@ set(cora_sources
     ${cora_source_dir}/commands/command_interpreter.cpp
     ${cora_source_dir}/commands/list_command.cpp
     ${cora_source_dir}/commands/play_command.cpp
+    ${cora_source_dir}/commands/search_command.cpp
     ${cora_source_dir}/commands/stop_command.cpp
     ${cora_source_dir}/logging/message_only_formatter.cpp
 )